First steps of the game
For now it has one the Score (which is for every alien that will be hit
the player will gain a score) which is linked with the player, Alien
generator ( which is suppose to work but there is a bug issue which is
while the aliens die no more spawns), Health which is allocated with the
Player controller ( When player hit the Health goes down but there is a
bug issue which is when the player ends up with ) Health points, the
points of health keeps going Under 0 for example -0 and the game doesn't
say game over). Still need images for the spaceship and aliens.